HUGE MERGE. Falcon coders, please read!
This changeset incorporates all of the remaining changes that differentiate the old codebase from WebUI v0.370. This includes a rewritten language string loading system, and a largely overhauled column handling system for the listviews. The rewritten language string loading allows for true lazy loading of language strings (so that en.js is never loaded even if the user uses a different language), and allows for dynamic changing of the interface language without requiring a full WebUI reload. The column handling overhaul ditches the use of magic numbers for reaching into each torrent job's data array in favor of named constants (CONST.TORRENT_* in constants.js). As well, new columns are simpler to add, and the default column configuration is trivial to change.

I've been holding off on the merge because I was trying to find a way to make the diff as small and readable as possible, but that proved to be a bit more difficult than was I had hoped (but considering the scope of the changes, I'd say it's hardly unexpected either). The string loading and column handling rewrites couldn't be separated because they are largely interdependent.

Anyway, the diff will be ugly -- sorry to those maintaining the Falcon branch :P Better than letting the codebase diverge too greatly. It felt like this was really starting to happen after seeing thomas' changeset #106 -- a hack to fix Directrix's use of magic numbers, which had already been (largely) resolved with the column handling rewrite.

@thomas: Just FYI, your changes won't immediately work with this code. If my changes prove to be as robust as I'm hoping, then all you'll need to do is to modify constants.js's CONST.TORRENT_* to reflect the new data fields added to /gui/?list=1, and it should Just Work™. :P